requirements.txt
datasets==2.20.0
langsmith==0.1.128
langchain_openai==0.2.0
langchain==0.3.0
langchain_community==0.3.0
langchain-huggingface==0.1.0
torch==2.4.0 #was 2.1.0 but venv gave error
termcolor==2.4.0
accelerate==0.33.0
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
einops==0.8.0
wheel==0.44.0
sentencepiece==0.2.0
protobuf==5.27.3
groq==0.10.0
matplotlib==3.9.2
seaborn==0.13.2
scipy==1.14.1
statsmodels==0.14.4
anthropic==0.40.0
together==1.3.14
google-genai==0.8.0
sentence-transformers==3.3.1
networkx==3.4.2
transformers==4.49.0
langchain-google-genai==2.1.3
smolagents==1.18.0
smolagents[e2b]==1.18.0
faiss-cpu==1.11.0

#faiss-gpu==1.7.2
#google-generativeai==0.8.4 #conflicting dependency with langchain-google-genai